refactor(components): add common dir
