+++ frontend/app/components/admin/dashboard/AdminPanel.tsx
@@ -1,10 +1,10 @@
 'use client';
 
+import StatusBadge from '@/app/components/common/StatusBadge';
 import type { QuizCard } from '@/app/types';
 
 import AdminLayout from '../layout/AdminLayout';
 import { statusColors } from '../shared/constants';
-import StatusBadge from '../shared/StatusBadge';
 
 interface AdminPanelProps {
   quizzes: QuizCard[];
