+++ frontend/app/components/admin/editor/index.ts
@@ -1,4 +1,3 @@
-export { default as ImageDropzone } from './ImageDropzone';
 export { default as QuestionListItem } from './QuestionListItem';
 export { default as QuestionPreview } from './QuestionPreview';
 export { default as QuizEditor } from './QuizEditor';
