+++ frontend/app/components/admin/shared/index.ts
@@ -1,2 +1,1 @@
 export { menuItems, statusColors, typeColors } from './constants';
-export { default as StatusBadge } from './StatusBadge';
