+++ frontend/app/components/common/ImageDropzone.tsx
