+++ frontend/app/components/common/ProgressBar.tsx
