+++ frontend/app/components/common/SliderTrack.tsx
