+++ frontend/app/components/common/index.ts
@@ -1,0 +1,5 @@
+export { default as ImageDropzone } from './ImageDropzone';
+export { default as ProgressBar } from './ProgressBar';
+export { default as SliderTrack } from './SliderTrack';
+export { default as StatusBadge } from './StatusBadge';
+export { default as SubmitButton } from './SubmitButton';
