+++ frontend/app/components/quiz/questions/RangeSlider.tsx
@@ -2,10 +2,11 @@
 
 import { useState } from 'react';
 
+import SliderTrack from '@/app/components/common/SliderTrack';
+import SubmitButton from '@/app/components/common/SubmitButton';
+
 import QuestionCard from '../shared/QuestionCard';
 import QuizLayout from '../shared/QuizLayout';
-import SliderTrack from '../shared/SliderTrack';
-import SubmitButton from '../shared/SubmitButton';
 
 interface RangeSliderProps {
   questionNumber: number;
